CSN: CSN-asimcore-304
Removed all tools made redundant by the 'awb' package and updated
the autoconf and make files.

Workspaces that have not been upgraded since the split of awb
will be broken by this checkin and you will need to manually
add <INSTALLDIR>/share/asim/awb/HEAD to the SEARCHPATH in the
awb.config file.